+++ src/mysql/upgrade/5.8.2-upgrade.php
@@ -15,6 +15,8 @@ require '../../Include/Config.php';*/
   
   unlink(SystemURLs::getDocumentRoot()."/FamilyList.php");
   unlink(SystemURLs::getDocumentRoot()."/PersonList.php");
+  unlink(SystemURLs::getDocumentRoot()."/sundayschool/SundaySchoolDashboard.php");
+  unlink(SystemURLs::getDocumentRoot()."/sundayschool/SundaySchoolClassView.php");
   
   $logger->info("End of delete :  all unusefull files");
 ?>
